+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changed:
 - Bug fix: in `bddOf` for `Dia` use `unsafeUpdate`
 - Bug fix in HasCacBDD-0.3.0.0: segfault on Apple M1 and ARM
 - Web: use `WEBPATH` for running behind non-rewriting proxy
+- CLI: only use colored text when stdout is a terminal
 
 ## v1.3.0 (2024-04-24)